# MAP-28: adds shards to SymbolAnimationCoordinatorMap