--- a/shared/test/CMakeLists.txt
+++ b/shared/test/CMakeLists.txt
@@ -10,6 +10,7 @@ FetchContent_MakeAvailable(Catch2)
 add_executable(tests 
   "TestActor.cpp"
   "TestGeoJsonParser.cpp"
+  "TestSymbolAnimationCoordinatorMap.cpp"
 )
 # Use mapscore _private_ include to allow testing functionality declared in internal headers.
 target_include_directories(tests PRIVATE