This PR adds Catch2 tests for the SymbolAnimationCoordinatorMap and fixes the test build for macOS.
With reproducible test cases, we now have a measurable benchmark. Additionally, I added shared ownership to the SymbolAnimationCoordinatorMap, which improves runtime performance.

before:

```
benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Benchmark random data                          100             1     4.15799 s 
                                        42.7119 ms     42.093 ms    44.6054 ms 
                                        4.98408 ms    1.84416 ms    10.8426 ms 
                                                                               
Simulate from log file                         100             1      21.509 s 
                                        211.488 ms    210.565 ms    212.385 ms 
                                        4.62717 ms    4.05888 ms    5.35695 ms 
```

after:

```
benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Benchmark random data                          100             1    940.766 ms 
                                        9.66332 ms    9.60304 ms    9.74142 ms 
                                        347.726 us    278.248 us    449.302 us 
                                                                               
Simulate from log file                         100             1     7.57967 s 
                                        74.5258 ms    72.4308 ms    77.2266 ms 
                                        12.0686 ms    9.43013 ms    16.1592 ms 
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **Chores**
	- Updated the `.gitignore` file to exclude the `cmake-build-test` directory.
	- Simplified platform-specific build settings in the `CMakeLists.txt` for better management of source files and include directories.
- **Tests**
	- Added comprehensive tests for the `SymbolAnimationCoordinatorMap`, including functionality and performance under concurrent conditions.
	- Updated GitHub Actions workflow to include benchmarks in the test execution.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->